Fixed (?) car collisions
Will do traffic lights next, that might fix the rest of my issues.
But, back car code isnt really working, probably because i cant combine stop values from the front and back of the car at the same time.
Need to ensure that when a car is on the crossing, it just keeps going instead of waiting for a pedestrian.